Complete refactor and remove redundant files.
