+++ that_depends/__init__.py
@@ -2,10 +2,9 @@
 
 from that_depends import providers
 from that_depends.container import BaseContainer
-from that_depends.entities.context import ContextScope, ContextScopes, get_current_scope
 from that_depends.injection import Provide, inject
 from that_depends.providers import container_context
-from that_depends.providers.context_resources import fetch_context_item
+from that_depends.providers.context_resources import ContextScope, ContextScopes, fetch_context_item, get_current_scope
 
 
 __all__ = [
